refactor: modularize extension for scalability
